backend/requirements.txt
Django==3.2.18
djangorestframework==3.14.0
django-filter==2.4.0
djoser==2.1.0
drf-extra-fields==3.7.0
gunicorn==20.0.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
reportlab==3.6.12
drf-yasg==1.21.7
pillow==10.3.0